Move package to src dir and rename to sdrun
The splitting of the MD-Molecules-Hoomd repository, means this is now
the statdyn-analysis repository. As the first part of this split, I am
reorganising the repository to include the module under the src
directory. This means that the code has to be installed correctly for it
to run at all. Since a significant portion of the functionality is
enabled through the command line interface, which does require
installation I think this is a reasonably sensible move.