Use molecule to create particle positions
This modifies how the particle positions for the additional particles of
a molecule are created. Rather than let the creation of the rigid bodies
define the creation, I am adding them upon initialisation, whether from
a file, with a crystal, or on a grid. This allows for more flexibility
like defining non-rigid molecules, or 'molecules' without bonds at all,
i.e. a binary mixture.