Add binary mixture molecule
This is a molecule that has no bonds between the particles, it is just a
binary mixture, defaulting to a binary mixture that is a good glass
former based on the Paper "Central role of thermal collective strain in
the relaxation of structure in a supercooled liquid" by Asaph
Widmer-Cooper and Peter Harrowell. The molecules are scaled up to be a
close match to the rest of the molecules I am using, with the large
particle having a $\sigma$ of 2 instead of 1.4, and the small having
a $\sigma$ of 1.43 instead of 1.