Tempfix - file system not releasing files in tests
The file system on quartz is not releasing the files being created in
the tests which results in the raising of an OSError and the test
failing. At the current point in time I have marked the tests as xfail,
so that it doesn't stop builds running, however I would like another
solution at some point.