[#2] - Only update snapshot with rank == 0
The issue was that when running with mpi, all the processes were trying
to update the particles. Hoomd expects that only the master process,
i.e. rank == 0 performs the update. To enforce this state, hoomd has a
comm module built in which can be used to query the current mpi rank,
which I have used to enforce the expected behaviour.