Update dependencies
Primarily this is pinning the quaternion dependency, and updating numpy
to 1.14.1. This is a result of the quaternion package building against
numpy 1.14.1 which means that is the version I have to use across the
board. I am fine with updating numpy since there are some features I
would like to look at, also the updated version of numpy_quaternion
doesn't have the distribution issues with pipenv I was running into
earlier.