Handle directory creation in parse_args function
This ensure that both the directory for the outfile, when it is
required, and the output directory are present. This fixes #1 and should
bring most of the directory creation errors to the beginning of the
simulation.

Additionally this changes the outfile_dir paramter in the
SimulationParams to output, to maintain consistency with the input
flags. This is something that is far less confusing from a maintenance
point of view. This is included in this commit since it is closely
intertwined.