Use molecule dimensions instead of crystal dimensions
The dimensions of the molecule are guaranteed to be present, as the
molecule is set to a default value of Trimer in the SimulationParams
class. Getting the parameter of the molecule also takes into account the
crystal parameters, since where the molecule is not overridden the
molecule of the crystal is returned.