Specify git version to checkout it meta.yaml
This specifies the version in both the package version and now
additionally in the source as the git_rev. This change now allows me to
go back to previous versions and build them, not just the current
version. Currently a point of concern since I am not sure the build and
upload is working completely on travis.