Make orthogonalisation of crystal explicit
The orthogonalisation of the crystal structure took place at the end of
the equilibration of the crystal structure. This has both hidden the
orthogonalisation step and means that it is always applied.

As part of the move towards calculating the free energy of crystal
structures I am moving towards having non-orthogonal crystals, which
previously wasn't possible.