Integrate harmonic force simulations into package
The harmonic force simulations are an important subset of the
simulations, enabling the calculation of the free energy of the crystal
structure. I have added all the logic to include these simulations as an
option from the command line. This means they should be as easy as any
other simulation type to perform.