Initialise simulation with SimulationParams
Rather than passing the molecule to the function for the initialisation
of a simulation I am instead passing the SimulationParams object,
allowing for more flexibility in the initialisation of the configuration
and ensuring more consistency across the initialisation functions.

Additionally since I am passing variables to the initialisation function
whose names exactly match those of the kwargs, I am just using the
positional arguments which should make the code simpler to read.