Merge snapshot minimization functions
There were two different functions for the minimization of snapshots,
with one being an NPH minimisation for quickly shrinking a simulation
cell from an initial configuration, and NVE for minimising a crystal
structure at a consistent pressure.

This creates a single minimisation function which combines the
functionality of both prior functions.