Create helper functions for the number of molecules
Getting the number of molecules in a simulation cell is an action I
perform fairly regularly, however this isn't entirely simple with many
edge cases and error checking in getting the right number. To make
everything simpler I have implemented a function which gets the number
of molecules and particles in a simulation cell and contains a number of
sanity checks.