Set integrator parameters for crystals properly
For the NPT crystal simulations the rescale_all parameter needs to be
set to True. This is not a parameter that exists for the NVT simulations
so needs to be set separately. Additionally the crystal parameters only
apply to the NPT simulations where the changes to the simulation cell
need to be adjusted for the crystals.