Making simulation orthorhombic is more explicit
An orthorhombic unit cell is no longer the default return state of
initialisation from a crystal. This paves the way for handling unit
cells that aren't orthorhombic. Additionally this makes the setting of a
unit cell as orthorhombic a more explicit action in the cases where this
is required and intended.