ENH Ensure unit cell dimensions are always len 3
The unit cell dimensions are now always of length 3, with the handling
of alternate values handled by the SimulationParams class. Where there
are missing values, sensible defaults, i.e. 1 are used.

This also updates the suite of tests for the SimulationParams ensuring
that the appropriate values are returned in addition to the correct
length.