ENH Use parameters to define integrator
This allows for more flexibility in the specification of the integrator
with the helper function providing all the logic for specifying the
required parameters. This means there is only a single place where the
integrator is defined.

The use of parameters makes it a lot clearer what is actually intended
in the integration, rather than trying to infer from the parameters
being set.