BUG Fix equil_interface simulations
The equilibrium equilibration was broken, running into issues with
particles being outside the simulation cell. The cause of this issue was
a block of code trying to be too fancy and create the simulation in some
circumstances. By removing that block of code the interface simulations
are now working again.