ENH Function to create interface from scratch
The creation of a liquid--crystal interface requires a number of steps
which can result in the simulation not working if not done
appropriately. The create_interface function is a helper function which
performs all the appropriate steps for the initial creation of the
interface. It takes into account both the create and melting steps of
the creation.

I have also included tests for this function ensuring stability in the
use of the function for creating and interface.