ENH Use master equilibrate function for create_interface
This is the start of integrating the master equilibrate function into
the rest of the simulations. This commit uses the equilibrate function
for the create_interface function. Additionally I have made the
create_interface function more robust, checking appropriate parameters
exist and also introducing some logic for the number of steps to run,
which should be fewer for the initialisation than for the melting part
of the creation.