ENH Use attrs for creating SimulationParams class
Attrs is a well developed project for generating classes with a range of
functionality. I have chosen attrs over Dataclasses since the design of
attrs is closer to how I am thinking about this class. By using attrs I
am able to focus on the bits of the class that include logic rather than
boilerplate.

The most notable feature of attrs is the idea that no python variables
are hidden and variables with a leading underscore has this stripped for
the keyword in the __init__ function. This is exactly the functionality
that I am looking for and is far simpler than the approach I had to take
using Dataclasses.

Attrs is also nice in that there is significantly more documentation
than Dataclasses which is not technically released yet as python 3.7 is
still in beta.