ENH Use Path consistently throughout package
I have defined the output, outfile, and infile variables in
SimulationParams to be a Path rather than a string. This is a
significantly more powerful representation of the path, allowing for
simple modification. This commit changes the useage throughout the
package to use Path, converting to a path when reading the arguments and
only converting back to a string when passing to hoomd functions which
don't support paths.