ENH Reimplement the argument parsing using click
This uses the click library to perform the argument parsing of the sdrun
command line tool. I have migrated to using click for it's simplicity of
understanding what the arguments to a function are, with the argument
parsing implemented right next to the functions they are used with.

This is not without constraints to the implementation of the command
line arguments, most notably being that the arguments for a group are
not passed along which does make for an tool that is a little harder to
understand at first. However there is the advantage of having
conceptually distinct options.

BREAKING CHANGES

This breaks the command line API, requiring a reordering of the
arguments when specifying the command. There are some conceptual
benefits to this approach, and a huge benefit with regards to the code
which is why I am allowing this change.