MAINT Remove individual equil_* functions
The equil_* functions were primarily rehashes of the same code, I
introduced an alternate function, equilibrate which takes the * part as
an argument resulting in less overlap of code.

I have now updated the package to use this updated function, in the
process removing all the code for the old functions.