ENH Make equil_type an option rather than argument
Having 3 arguments for the equil option was a little too confusing to
explain and also leading to bugs since I can't skip one of the options.
Since most simulations are going to have the equil_type of liquid, I am
using this as the default option with any of the other options
specifiable using the --equil-type options.