ENH Remove gradient when equilibrating crystal
Equilibrating the crystal lattice occurs at a single temperature which
is given by the init_temp variable. Since I temporarily set the
temperature to init_temp, there is a temperature gradient from init_temp
to init_temp. So this sets init_temp to none for this section of the
equilibration.