ENH Add maximum number of steps for temperature ramping
This caps the maximum number of steps for the ramping of the temperature
at 750k. This is the value that would have been set for a million steps.
This change should make the longer equilibrations somewhat faster, and
equilibrating any slower shouldn't result in change in the simulation.
This is most important for higher temperatures where quickly ramping up
the energy can cause issues.