TST Extend test regime of sdrun_test and remove duplicates
Running the click cli runner runs the complete command passed. This
has meant that the tests of running the command have been running twice.
I have removed the click set of testing to reduce overlapping tests.

The other change I have made is to increase the area of the commands
tested. Now I am testing all the crystal structures for a longer
simulation period which is to cover as much of the command area for the
MPI tests as possible.