BUG Only load HarmonicForceCompute when used
The HarmonicForceCompute module has a glibc dependency that is not found
on all systems I am using it on for some reason. To rectify this I am
only importing the module when the set_harmonic_force function is used.
This is the only place in the codebase where this module is used.