Extract group from SimulationParams class (#15)
This completely removes the hoomd group from the SimulationParams class,
instead passing it to functions that require it.

The group is something that is specific to the simulation being run
rather than a more general thing. This means it doesn't fit into the
SimulationParams class. I was previously using the class as a carrier,
although that didn't make sense. This commit just removes this attribute
entirely.