MAINT Move z2quaternion function to util module
This moves the z2quaternion function to the util module, since it is a
utility function. This has been problematic previously with the typing
of the params module introducing circular dependencies. I have now
worked out how to resolve the circularity of the dependencies, and have
so moved the function.