ENV Include rowan for quaternion maths
Rowan is a python package for performing quaternion operations. It is
super fast at the types of operations I am using in my simulations, so
a useful addition.

Eventually this means I can replace the quaternion functions I have
written with a more powerful and better tested and documented set of
functions.