MAINT Use init_from_crystal to initialise from none
The init_from_none function was a lot of code which duplicated the
init_from_crystal function. In making the crystal class more general I
have also made it significantly simpler to create an arbitrary instance
of the class. I have used this to create a crystal which is then
initialised in the same way as the crystal.