BUG Set default unit cell position to the origin
With the mangling that occurs when the particle positions are shifted to
center the simulation cell, it turns out the best place to put particles
is at the origin, rather than in the center of the cell. This sets the
default position to the center of the unit cell and changes the
SquareCircle and CubicSphere crystals back.