BUG Fix molecule equality comparisons
In the comparison of the molecules, there was an issue in attempting to
compare numpy arrays, which require an np.all or np.any. Since this was
in a function generated by attrs I have defined these functions myself,
including additional comparisons for the subclasses. These comparisons
all have tests.