ENH Ensure the NumBodies class always returns an int
The `hoomd.create_snapshot` function only accepts a python `int` as the
argument for the number of particles. The origin of the issues was the
NumBodies class having a value of type `np.int32`. This commit adds a
converter to the initialisation of the NumBodies class which ensures the
stored values are always a python int.