ENH Use hoomd to randomise momenta and thermalise
This changes the implementation of the randomise momenta function to use
hoomd for the randomisation, through the `integrator.randomize_velocities`
function. This is a significantly more robust implementation than the
one I have been using. Additionally since it uses the temperature for
the randomisation I am able to use this same function for the
thermalisation of a configuration.

In changing the randomise_momentum function I have moved it from the
util module to the initialise module since it conceptually fits much
better.