BUG Proper testing of simulation thermodynamics
Even though I had the code to test the thermodynamics of a simulation,
the tests were not actually running because of the way pytest does the
test setup and discovery. To make it simpler for testing I created a new
test file which doesn't take so long to run, adapting the class to a
series of functions for testing. This involved the creation of a
SimStatus class holding a number of the simulation parameters. This
class will likely make it's way into the main sdrun codebase.