statdyn-simulation 0.6.0
Versions prior to this release have a significant bug in the handling of
rigid molecules. All molecules defined previously have a mass of 1,
instead of the intended mass of 3, while the moment-of-inertia is
calculated for a molecule of mass 3.

This version introduces significant breaking changes, including losing
the ability to read input/output files from previous versions. This is
the result of significant bug in the definition of rigid molecules which
required a rewrite of the molecule handling.

Because of the significance of the bug, and that I now need to start
collecting two years worth of data again, the non-essential items have
been left out to get this done quickly. I do intend to build in support
for the older configurations in a future release.

- Molecules have additional non-interacting central rigid body
- Use the shifted lennard jones potential instead of standard LJ

- Rewrite of molecule class to fix bug
- Equality operators that work for Molecule and Crystal classes
- Ensure NumBodies always returns a python int
- Util function to compute the kinetic energy of a snapshot
- Use hoomd functions to randomise momenta
- Simplify the __repr__ of classes
- Use matrix algebra to generalise computation of quantities

- Add explicit testing for masses
- Improve test fixtures making them global in conftest.py
- Add and improve tests for crystal positions
- Specifc testing of rigid body masses
- Include testing of thermodynamic properties
- Test the initialisation from a file
- Update test initial configurations

- Update to hoomd 2.3.0
- Include rowan for the computation of quaternion values
- Add mpi4py to manage mpi state
- Install all tools with conda

- Include more pylint tests