MAINT Remove harmonic_force from repo
The harmonic force is turning out to be rather problematic. I don't
actually trust it's ability to generate a harmonic potential on a
rotation and it is also just annoying to maintain compatibility.
Since I haven't worked out a docker container for it, I am going with
the approach of just removing the capability.