BUG Properly test dynamics flag on command line
The flags for activating the strings were the wrong way around, which
made the activation/deactivation difficult. In addition to this issue,
and the lack of a default value, the tests which I thought would catch
the issue didn't actually test the issue.

This commit fixes the tests ensuring that they don't work when things
are broken, rather than just working all the time.