BUG None value for lattice-lengths not begin recognised
The None value for the lattice lengths was being set to an empty tuple
which is not a None value and hence was being set with the
initialisation of the SimulationParams class in the main function.

This commit adds additional testing and comments which formed the
debugging process for finding this bug. This additionally tests that the
bug will not reoccur.