docs: Make documentation work again
This updates the documentation to cope with the updated project
structure. This has involved many changes to the structure, updating the
files and modules that are pointed to.

This includes updates to the environment and the makefile including the
required packages for the compilation of the documentation in addition
to a makefile rule for building the documentation.