build: Use setup.py for installing packages
Pipenv just isn't working for me, unable to resolve dependencies and is
just at the moment more hassle than it is worth. Instead I am using the
setup.py to install packages, which I can get around to using
requirements.txt files.

This updates the dockerfile as well so as to enable to build to
complete.