refactor: Replace assertions with raising ValueErrors
Raising ValueErrors is the more correct approach to checking errors,
which will not be removed when running using certain flags. Additionally
the value error is more suited to being handled by higher functions.

In writing these new checks there is also the information required to
fix the issue, that is, the value which has caused the error is printed
along with some instructions for what types of values are supported.

Ref #28