feat: Replace assertion checks with ValueErrors
This makes the error handling more concise, enabling different types of
errors to reflect more user issues in the case of ValueErrors, or code
issues in the case of RuntimeErrors. Using the exceptions is more
standard practice, and along with the move to the exceptions I have
made the error messages from them more useful, giving reference to what
needs to be fixed and the value which caused the error in the first
place.

Ref #28