fix: Randomise momenta from iteration_id
The iteration id is a property which is intended to be used as a seed
for a random number. Well at the very least have a different sequence of
random numbers for each value. This commit ensures that random numbers
are generated upon initialisation even when there is only an iteration
id specified.